Update steam-workshop.md
When I tried to upload my first mod with a Cloud Thumbnail, I received the "Mod is being Uploaded" message and then nothing else for an hour.  Only after digging around in the Forums did I find the suggestion of designating the uploaded thumbnail file as Local instead of Cloud.  When I did that, the mod uploaded immediately.  Thank you for considering adding this to the doc.  Z