Suggest linking to your mod from BoardGameGeek
BoardGameGeek contains a wealth of knowledge about board games, and it would be great if it was easier to get from there to places where you can play them.

Incidentally, I’ve written a [little webpage](https://www.mavit.org.uk/boardgames-online/) to help with that, but of course it’s only useful if the data is in BoardGameGeek in the first place.